CNDB-11518: Split o.a.c.index.sai.cql.VectorTypeTest to prevent timeouts
