+++ test/unit/org/apache/cassandra/index/sai/cql/VectorTester.java
@@ -22,9 +22,14 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.List;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
 import org.apache.commons.lang3.reflect.FieldUtils;
 import org.junit.Before;
+import org.junit.Ignore;
+import org.junit.runner.RunWith;
+import org.junit.runners.Parameterized;
 
 import io.github.jbellis.jvector.graph.GraphIndexBuilder;
 import io.github.jbellis.jvector.graph.GraphSearcher;
@@ -37,6 +42,8 @@ import org.apache.cassandra.db.ColumnFamilyStore;
 import org.apache.cassandra.db.Keyspace;
 import org.apache.cassandra.index.sai.IndexContext;
 import org.apache.cassandra.index.sai.SAITester;
+import org.apache.cassandra.index.sai.SAIUtil;
+import org.apache.cassandra.index.sai.disk.format.Version;
 import org.apache.cassandra.index.sai.disk.v2.V2VectorIndexSearcher;
 import org.apache.cassandra.index.sai.disk.v5.V5VectorPostingsWriter;
 import org.apache.cassandra.index.sai.disk.vector.ConcurrentVectorValues;
@@ -163,5 +170,47 @@ public class VectorTester extends SAITester
         }
 
         return matches * 1.0 / result.size();
+    }
+
+    /**
+     * {@link VectorTester} parameterized for {@link Version#CA} and {@link Version#DC}.
+     */
+    @Ignore
+    @RunWith(Parameterized.class)
+    abstract static class Versioned extends VectorTester
+    {
+        @Parameterized.Parameter
+        public Version version;
+
+        @Parameterized.Parameters(name = "{0}")
+        public static Collection<Object[]> data()
+        {
+            return Stream.of(Version.CA, Version.DC).map(v -> new Object[]{ v }).collect(Collectors.toList());
+        }
+
+        @Before
+        @Override
+        public void setup() throws Throwable
+        {
+            super.setup();
+            SAIUtil.setLatestVersion(version);
+        }
+    }
+
+    abstract static class VersionedWithChecksums extends Versioned
+    {
+        @Override
+        public void flush()
+        {
+            super.flush();
+            verifyChecksum();
+        }
+
+        @Override
+        public void compact()
+        {
+            super.compact();
+            verifyChecksum();
+        }
     }
 }
